Improve parsing of list-object combinations
Use bounded lookahead to improve parsing of lists of objects and empty
keys that should parse as null.

Previously an empty key would parse as an object that contained keys
that should have been siblings. Includes an expanded test of null parsing
and a hie.yaml file that occurred in the wild and did not parse correctly.